Fixing fucked up errors
